+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 All notable changes to `pmatseykanets/laravel-hipchat-slash` will be documented in this file
 
 ## [Unreleased]
+### Added
+- Configurable search config
 
 ## [0.3.0](https://github.com/pmatseykanets/laravel-scout-postgres/releases/tag/v0.3.0) - 2017-01-04
 ### Changed